Fixed tests and update docs
