+++ docs/clients/s3.md
@@ -106,12 +106,25 @@ foreach ($result->getBody()->getChunks() as $chunk) {
 
 ## Virtual Hosted-Style Requests
 
-When calling AWS endpoints, Async Aws uses [Virtual Hosted-Style Requests](https://docs.aws.amazon.com/AmazonS3/latest/dev/VirtualHosting.html):
-The bucket name is part of the endpoint's host. To disable this behavior, use
-the parameter `pathStyleEndpoint` when initializing the client.
+When calling AWS endpoints, AsyncAws uses [Virtual Hosted-Style Requests](https://docs.aws.amazon.com/AmazonS3/latest/dev/VirtualHosting.html):
+The bucket name is part of the endpoint's host. To change this behavior, and use
+"path styled endpoints" instead, set `pathStyleEndpoint` parameter to `true` when
+initializing the client.
 
 ```php
 use AsyncAws\S3\S3Client;
 
 $s3 = new S3Client(['pathStyleEndpoint' => true]);
+```
+
+## Non-AWS S3 endpoints
+
+To use the `S3Client` with example Digital Oceans' Spaces, you need to initialize
+the `S3Client` with your endpoint.
+
+```php
+$s3 = new S3Client([
+    'endpoint' => 'https://fra1.digitaloceanspaces.com',
+    'pathStyleEndpoint' => true,
+]);
 ```
